requirements.txt
# 基本依赖
# 支持Python 3.12

# Web框架
flask>=2.0.0

# 数据处理
pandas>=1.3.0
numpy>=1.20.0

# API请求
requests>=2.25.0

# 自然语言处理
nltk>=3.6.0
spacy>=3.1.0
python-dotenv
transformers>=4.8.0
torch>=1.8.0
# spaCy模型将通过命令行安装

# 数据可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 文档解析
pyPDF2>=1.26.0
biopython>=1.79

# Web应用会话管理
flask-session>=0.4.0

# 数据库
# sqlite3 是Python标准库，不需要通过pip安装

# 其他工具
tqdm>=4.61.0